cleanup: consolidate kernel-annotations-plugin and minor fixes
Move duplicate kernel-annotations-plugin from both apps into a shared
version in @calab/ui/chart with a toMs multiplier for unit flexibility.
Also fix seed phase double .map() to single-pass and remove unnecessary
createMemo for static plugin array in catune KernelDisplay.
